concerto: reshape PortfolioWindow into WavesView; drop RepoSidebarWindow
Reshape the proven PortfolioWindow into the target shape (rename -> WavesView):
burgundy repo rail + wave list (WaveRow) + new-wave launcher, and clicking a wave
attaches its /goal agent in an embedded tmux terminal. Reuses WaveSidebar's
Color.loopflowBurgundy treatment, WaveRow, CreateWaveSheet, and the
TerminalWorkspaceView attach path -- no NavigationSplitView, no rebuilt styling.
Deletes the RepoSidebarWindow sketch + old dashboard cards. Net -270 lines.
